Added CSS & correct images
I’ve had trouble linking local files to the project(.js, .css, images)
after some testing I found out it only works if ur linking to a file
from the web. There probably is a real solution to this but nothing
worked for me. I uploaded the correct .SVG file on to a site and have
the login and register page link to it for the background.